fdsfs

+++ twbot/bot.py
@@ -811,34 +811,37 @@ class InstaBot:
                     number_class.ban_number(self.phone_number,country_code,china=china)
                     continue
                 
-                if self.find_element('Incorrect number','//android.view.View[@content-desc="Looks like your mobile number may be incorrect. Try entering your full number, including the country code."]'):
-                    number_class.ban_number(self.phone_number,country_code,china=china)
-                    continue
+                otp_page = self.find_element('Confirmation code input','//android.view.View[@content-desc="Enter the confirmation code"]',page='OTP page')
+                if not otp_page :
                 
-                elif self.click_element('Create account','//android.widget.Button[@content-desc="Create new account"]'):
-                    ...
+                    if self.find_element('Incorrect number','//android.view.View[@content-desc="Looks like your mobile number may be incorrect. Try entering your full number, including the country code."]'):
+                        number_class.ban_number(self.phone_number,country_code,china=china)
+                        continue
                     
-                elif self.find_element('otp page', '//android.view.View[@text="Enter the confirmation code"]',timeout=5):
-                    ...
-                    
-                elif self.find_element('something went wrong','//android.view.View[@content-desc="Something went wrong. Please try again later."]',timeout=2):
-                    number_class.ban_number(self.phone_number,country_code,china=china)
-                    self.df.loc['avd']=self.emulator_name
-                    self.df.to_csv('delete_avd.csv', index=False)
-                    print(f'add this {self.emulator_name} avd in delete local avd list')
-                    self.user_avd.delete()
-                    return 'delete_avd' ,''
+                    elif self.click_element('Create account','//android.widget.Button[@content-desc="Create new account"]'):
+                        ...
+                        
+                    elif self.find_element('otp page', '//android.view.View[@text="Enter the confirmation code"]',timeout=5):
+                        ...
+                        
+                    elif self.find_element('something went wrong','//android.view.View[@content-desc="Something went wrong. Please try again later."]',timeout=2):
+                        number_class.ban_number(self.phone_number,country_code,china=china)
+                        self.df.loc['avd']=self.emulator_name
+                        self.df.to_csv('delete_avd.csv', index=False)
+                        print(f'add this {self.emulator_name} avd in delete local avd list')
+                        self.user_avd.delete()
+                        return 'delete_avd' ,''
 
-                elif self.find_element('phone number page', '''//android.view.View[@text="What's your mobile number?"]''',timeout=2):
-                    self.df.loc['avd']=self.emulator_name
-                    self.df.to_csv('delete_avd.csv', index=False)
-                    print(f'add this {self.emulator_name} avd in delete local avd list')
-                    self.user_avd.delete()
-                    return 'delete_avd'  ,''    
-                
-                elif self.find_element('name page', '''//android.view.View[@text="What's your name?"]''',timeout=2):
-                    return True ,''
-                
+                    elif self.find_element('phone number page', '''//android.view.View[@text="What's your mobile number?"]''',timeout=2):
+                        self.df.loc['avd']=self.emulator_name
+                        self.df.to_csv('delete_avd.csv', index=False)
+                        print(f'add this {self.emulator_name} avd in delete local avd list')
+                        self.user_avd.delete()
+                        return 'delete_avd'  ,''    
+                    
+                    elif self.find_element('name page', '''//android.view.View[@text="What's your name?"]''',timeout=2):
+                        return True ,''
+                    
                 for I_otp in range(10) :
                     
                     otp = number_class.get_sms(self.phone_number,country_code,china=china)
